requirements.txt
adbutils==0.13.1
aircv==1.4.6
allure-pytest==2.9.45
allure-python-commons==2.9.45
apkutils2==1.0.0
Appium-Python-Client==2.1.2
async-generator==1.10
attrs==21.4.0
cached-property==1.5.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cigam==0.0.3
colored==1.4.3
cryptography==36.0.1
decorator==5.1.1
Deprecated==1.2.13
deprecation==2.1.0
EasyProcess==1.1
entrypoint2==1.0
enum34==1.1.10
facebook-wda==1.4.6
filelock==3.4.2
h11==0.13.0
idna==3.3
iniconfig==1.1.1
logzero==1.7.0
lxml==4.7.1
macaca==0.1.0
mss==6.1.0
numpy==1.22.1
opencv-python==4.5.5.62
outcome==1.1.0
packaging==20.9
Pillow==9.0.0
pluggy==1.0.0
progress==1.6
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pyelftools==0.27
PyMySQL==1.0.2
pyOpenSSL==21.0.0
pyparsing==3.0.7
pyscreenshot==3.0
pytest==6.2.5
pytest-assume==2.4.3
pytest-ordering==0.6
requests==2.27.1
retry==0.9.2
retrying==1.3.3
selenium==4.1.0
simple-tornado==0.2.2
simplejson==3.17.6
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
tidevice==0.6.1
toml==0.10.2
tornado==6.1
trio==0.19.0
trio-websocket==0.9.2
uiautomator2==2.16.13
urllib3==1.26.8
wd==1.0.1
weditor==0.6.5
whichcraft==0.6.1
wrapt==1.13.3
wsproto==1.0.0
xlrd==2.0.1
XlsxWriter==3.0.2
xmltodict==0.12.0
